separate fpu and sfpu node
