+++ tt_metal/tt-llk/tests/python_tests/fuser/blackhole/fpu/reduce_block_max_runtime.py
@@ -5,7 +5,7 @@
 from typing import List
 
 from fuser.block_data import BlockData
-from fuser.compute_node import ComputeNode
+from fuser.fpu_node import FpuNode
 from fuser.fused_loop import FusedLoop, LoopBlockRow
 from fuser.fused_operation import FusedOperation
 from fuser.fuser_config import GlobalConfig
@@ -26,7 +26,7 @@ class ReduceBlockMaxRuntimeFpu(ReduceBlockMaxFpu):
         self,
         operation: FusedOperation,
         config: GlobalConfig,
-        compute_unit: ComputeNode,
+        compute_unit: FpuNode,
         block: BlockData,
     ) -> str:
         ct_dim = block.block_tiles_x
@@ -37,7 +37,7 @@ class ReduceBlockMaxRuntimeFpu(ReduceBlockMaxFpu):
         self,
         operation: FusedOperation,
         config: GlobalConfig,
-        compute_unit: ComputeNode,
+        compute_unit: FpuNode,
         block: BlockData,
     ) -> str:
         dest_acc = config.dest_acc.cpp_enum_value
@@ -47,7 +47,7 @@ class ReduceBlockMaxRuntimeFpu(ReduceBlockMaxFpu):
         self,
         operation: FusedOperation,
         config: GlobalConfig,
-        compute_unit: ComputeNode,
+        compute_unit: FpuNode,
         block: BlockData,
     ) -> str:
         return "_llk_math_reduce_block_max_row_uninit_runtime_();\n"
