+++ tt_metal/tt-llk/tests/python_tests/fuser/wormhole/fpu/matmul_no_mop.py
@@ -5,8 +5,8 @@
 from typing import List
 
 from fuser.block_data import BlockData
+from fuser.fpu_node import FpuNode
 from fuser.fused_loop import FusedLoop, LoopBlock
-from fuser.fused_math import ComputeNode
 from fuser.fused_operation import FusedOperation
 from fuser.fuser_config import GlobalConfig
 
@@ -28,7 +28,7 @@ class MatmulNoMopFpu(MatmulFpu):
         self,
         operation: FusedOperation,
         config: GlobalConfig,
-        compute_unit: ComputeNode,
+        compute_unit: FpuNode,
         block: BlockData,
     ) -> str:
         stage = operation.stage_id
@@ -54,7 +54,7 @@ class MatmulNoMopFpu(MatmulFpu):
         self,
         operation: FusedOperation,
         config: GlobalConfig,
-        compute_unit: ComputeNode,
+        compute_unit: FpuNode,
         block: BlockData,
     ) -> str:
         math_fidelity = compute_unit.math_fidelity.cpp_enum_value
@@ -74,7 +74,7 @@ class MatmulNoMopFpu(MatmulFpu):
         self,
         operation: FusedOperation,
         config: GlobalConfig,
-        compute_unit: ComputeNode,
+        compute_unit: FpuNode,
         block: BlockData,
     ) -> str:
         return "_llk_math_matmul_uninit_no_mop_();\n"
